feat: created component Nav
