+++ src/components/Home/HomeHeader/HomeHeader.scss
@@ -2,10 +2,6 @@
 
 // MOBILE FIRST
 .chazz-header {
-  // display: flex;
-  // flex-direction: row;
-  // align-content: space-between;
-  //
   background-image: url("video.png");
   background-repeat: no-repeat;
   background-position: -330px -100px;
@@ -17,131 +13,7 @@
     height: 100vh;
   }
 
-  .chazz-logo {
-    width: 137px;
-    display: block;
-    float: left;
-    position: relative;
-    left: 4%;
-    top: 30px;
-  }
-
-  .nav {
-    display: none;
-
-    a:hover{
-        color: var(--black);
-    }
-  }
-
-
-
-  .burger-menu{
-    width: 100%;
-    padding: 36px 16px 16px 16px;
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: flex-end;
-    align-items: center;
-    background-color: transparent;
-    transition: all .5s ease;
-    position: fixed;
-    z-index: 10;
-    &.active{
-        background-color: var(--black);
-    }
-
-    input#open-close { 
-        visibility:hidden;
-        position: absolute;
-        top: -9999px;
-    }
-    
-    .toggle-button {
-        display:block;
-        width:26px;
-        height:26px;
-        position:relative;
-        cursor: pointer;
-        box-sizing: border-box;
-    }
-
-    .toggle-button:before, .toggle-button:after {
-        position:absolute;
-        display:block;
-        content:" ";
-        width: 100%;               
-        box-sizing: border-box;
-        // left: calc(100% / 4);
-        transition: all 0.2s ease-out;
-    }
-
-    .toggle-button:before {
-        top: calc(100% / 3);
-        border-radius: 100px;
-        background-color:var(--white);
-        border-top:2px solid var(--white);
-        background-color: transparent;
-    }
-
-    .toggle-button:after {
-        border-radius: 100px;
-        background-color:var(--white);
-        border-top:2px solid var(--white);
-        bottom:calc(100% / 4);
-        width: 65%;
-        right: 0;
-    }
-
-    input#open-close:checked + .toggle-button:before, input#open-close:checked + .toggle-button:after {
-        top:calc(100% / 2);
-        height:1px;
-        border-bottom:0;
-    }
-
-    input#open-close:checked + .toggle-button:before {
-        transform: rotate(45deg);
-        width: 100%; 
-    }
-
-    input#open-close:checked + .toggle-button:after {
-        transform: rotate(-45deg);
-        width: 100%;
-    }
-
-    .burger-nav{
-        width: 100%;
-        display: flex;
-        flex-flow: column nowrap;
-        justify-content: center;
-        align-items: flex-start;
-        height: 0vh;
-        opacity: 0;
-        overflow: hidden;
-        transition: all .5s ease;
-        &.active{
-            height: 100vh;
-            opacity: 1;
-        }
-    }
-    .burger-ul{
-        list-style: none;
-        flex-direction: column;
-        padding-left: 0;
-    }
-
-    .burger-a{
-        text-decoration: none;
-        color: #FFAFC4;
-        font-weight: 500;
-        font-size: 48px;
-        line-height: 55.2px;
-
-        &:hover{
-            color: var(--white);
-        }
-    }
-  }
+  
 }
 
 .chazz-title {
@@ -179,7 +51,6 @@
 
   p {
     margin: 0;
-    // margin-top: 12px;
   }
 
   button {
@@ -203,7 +74,6 @@
   .chazz-cookies {
     display: none;
     left: 4%;
-    // height: 25%;
 
     button {
       padding: 1% 25% 1% 25%;
@@ -215,7 +85,7 @@
 @media (min-width: 425px) {
   .chazz-cookies {
     display: none;
-    // top: -14rem;
+
     button {
       padding: 3% 31% 3% 31%;
       margin-top: 23px;
@@ -232,63 +102,6 @@
 
     .chazz-header{
         background-position: 30% -100px;
-        .nav {
-            margin-top: 1.5%;
-            display: flex;
-            float: right;
-            padding-left: 0;
-        }
-        .nav ul {
-            display: flex;
-        }
-
-        .nav ul li {
-            list-style-type: none;
-            margin-right: 30px;
-        }
-
-        .nav ul li a {
-            text-decoration: none;
-            color: var(--white);
-        
-            &:hover {
-                text-decoration: none;
-            }
-        }
-
-
-        .burger-menu{
-            display: none;
-        }
-    }
-
-    .nav {
-      font-size: 1rem;
-      margin-top: 3.5%;
-      display: flex;
-      float: right;
-      padding-left: 0;
-      margin-right: 1%;
-    }
-    .nav ul {
-      display: flex;
-    }
-
-    .nav ul li {
-      list-style-type: none;
-      margin-right: 30px;
-    }
-
-    .nav ul li a {
-      text-decoration: none;
-      color: var(--white);
-      font-weight: 500;
-
-      &:hover {
-        text-decoration: none;
-        color: var(--black);
-      }
-    }
   }
 
   .chazz-title {
@@ -331,14 +144,7 @@
   }
 
 
-@media (min-width: 1024px) {
-  .chazz-header {
-    .nav {
-      margin-top: 2.5%;
-      font-size: 19px;
-    }
-  }
-}
+
 
 @media (min-width: 1280px) {
   .chazz-header {
@@ -346,18 +152,6 @@
     height: 100vh;
     background-position: 0 0%;
 
-    // .nav {
-    //     // margin-left: 0;
-    // }
-
-    .chazz-logo {
-      width: 15%;
-      top: 40px;
-    }
-
-    .nav ul li {
-      margin-right: 36px;
-    }
   }
 
   .chazz-title {
@@ -396,7 +190,6 @@
 
   .chazz-title {
     top: 40%;
-    // left: 85px;
   }
 }
 
@@ -421,4 +214,5 @@
       padding: 0.5% 1% 0.5% 1%;
     }
   }
+}
 }
