+++ src/components/Nav/Nav.scss
@@ -1,0 +1,213 @@
+@import "./src/styles.scss";
+
+// MOBILE FIRST
+
+.chazz-logo {
+    width: 137px;
+    display: block;
+    float: left;
+    position: relative;
+    left: 14px;
+    top: 40px;
+    z-index: 99;
+  }
+
+  .nav {
+    display: none !important;
+
+    a:hover{
+        color: var(--black);
+    }
+  }
+
+
+
+  .burger-menu{
+    width: 100%;
+    padding: 36px 16px 16px 16px;
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: flex-end;
+    align-items: center;
+    background-color: transparent;
+    transition: all .5s ease;
+    position: absolute;
+    z-index: 10;
+    &.active{
+        background-color: var(--black);
+    }
+
+    input#open-close { 
+        visibility:hidden;
+        position: absolute;
+        top: -9999px;
+    }
+    
+    .toggle-button {
+        display:block;
+        width:26px;
+        height:26px;
+        position:relative;
+        cursor: pointer;
+        box-sizing: border-box;
+    }
+
+    .toggle-button:before, .toggle-button:after {
+        position:absolute;
+        display:block;
+        content:" ";
+        width: 100%;               
+        box-sizing: border-box;
+        transition: all 0.2s ease-out;
+    }
+
+    .toggle-button:before {
+        top: calc(100% / 3);
+        border-radius: 100px;
+        background-color:var(--white);
+        border-top:2px solid var(--white);
+        background-color: transparent;
+    }
+
+    .toggle-button:after {
+        border-radius: 100px;
+        background-color:var(--white);
+        border-top:2px solid var(--white);
+        bottom:calc(100% / 4);
+        width: 65%;
+        right: 0;
+    }
+
+    input#open-close:checked + .toggle-button:before, input#open-close:checked + .toggle-button:after {
+        top:calc(100% / 2);
+        height:1px;
+        border-bottom:0;
+    }
+
+    input#open-close:checked + .toggle-button:before {
+        transform: rotate(45deg);
+        width: 100%; 
+    }
+
+    input#open-close:checked + .toggle-button:after {
+        transform: rotate(-45deg);
+        width: 100%;
+    }
+
+    .burger-nav{
+        width: 100%;
+        display: flex;
+        flex-flow: column nowrap;
+        justify-content: center;
+        align-items: flex-start;
+        height: 0vh;
+        opacity: 0;
+        overflow: hidden;
+        transition: all .5s ease;
+        &.active{
+            height: 100vh;
+            opacity: 1;
+        }
+    }
+    .burger-ul{
+        list-style: none;
+        flex-direction: column;
+        padding-left: 0;
+    }
+
+    .burger-a{
+        text-decoration: none;
+        color: var(--white);
+        font-weight: 500;
+        font-size: 48px;
+        line-height: 55.2px;
+
+        &:hover{
+            color: #FFAFC4;
+        }
+    }
+  }
+
+
+  // MEDIA QUERIES
+
+  @media (min-width: 768px) {
+
+    .chazz-logo{
+        width: 162px;
+        top: 30px;
+        left: 49px
+    }
+
+    .nav {
+        margin-top: 1.5%;
+        display: flex !important;
+        float: right;
+        padding-left: 0;
+    }
+    .nav ul {
+        display: flex;
+    }
+
+    .nav ul li {
+        list-style-type: none;
+        margin-right: 30px;
+    }
+
+    .nav ul li a {
+        text-decoration: none;
+        color: var(--white);
+    
+        &:hover {
+            text-decoration: none;
+        }
+    }
+
+
+    .burger-menu{
+        display: none;
+    }
+
+    .nav {
+      font-size: 1rem;
+      margin-top: 3.5%;
+      display: flex;
+      float: right;
+      padding-left: 0;
+      margin-right: 1%;
+    }
+    .nav ul {
+      display: flex;
+    }
+
+    .nav ul li {
+      list-style-type: none;
+      margin-right: 30px;
+    }
+
+    .nav ul li a {
+      text-decoration: none;
+      color: var(--white);
+      font-weight: 500;
+
+      &:hover {
+        text-decoration: none;
+        color: var(--black);
+      }
+    }
+  }
+
+
+  @media (min-width: 1024px) {
+      .nav {
+        margin-top: 2.5%;
+        font-size: 19px;
+      }
+  }
+
+  @media (min-width: 1280px) {
+  
+      .nav ul li {
+        margin-right: 36px;
+      }
+    }
